add comment exceptoion

+++ src/main/java/com/jzy/community/controller/PublishController.java
@@ -27,7 +27,7 @@ public class PublishController {
     private QuestionService questionService;
 
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id") Integer id,
+    public String edit(@PathVariable(name = "id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
@@ -49,7 +49,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description" ,required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value = "id") Integer id,
+            @RequestParam(value = "id") Long id,
             HttpServletRequest request,
             Model model) {
         model.addAttribute("title",title);
